feat: 좋아요 쿠키 기능 구현
- 이미 좋아요를 누른 사람은, 좋아요를 누를 수 없도록 쿠키를 사용해 구현함

+++ backend/src/main/java/feedzupzup/backend/feedback/domain/LikeFeedbacks.java
@@ -6,17 +6,17 @@ import java.util.Set;
 
 public class LikeFeedbacks {
 
-    private final Set<Integer> likeFeedbacks = new HashSet<>();
+    private final Set<Long> likeFeedbacks = new HashSet<>();
 
-    public void add(final int value) {
+    public void add(final long value) {
         likeFeedbacks.add(value);
     }
 
-    public boolean hasFeedbackId(final int value) {
+    public boolean hasFeedbackId(final long value) {
         return likeFeedbacks.contains(value);
     }
 
-    public Set<Integer> getLikeFeedbacks() {
+    public Set<Long> getLikeFeedbacks() {
         return Collections.unmodifiableSet(likeFeedbacks);
     }
 }
